ref(rust_bindings): add `from_raw()` functions
